refactor: use int for token lengths and drop redundant name lengths

+++ include/oak_token.h
@@ -85,9 +85,8 @@ struct oak_token_t
   int line;
   int column;
   int offset;
-  /* Lexeme length in bytes; same unsigned domain as `oak_string` / name
-   * lengths. */
-  usize length;
+  /* Lexeme length in bytes. */
+  int length;
   char text[];
 };
 
@@ -95,7 +94,7 @@ OAK_API enum oak_token_kind_t oak_token_kind(const struct oak_token_t* token);
 OAK_API int oak_token_line(const struct oak_token_t* token);
 OAK_API int oak_token_column(const struct oak_token_t* token);
 OAK_API int oak_token_offset(const struct oak_token_t* token);
-OAK_API usize oak_token_size(const struct oak_token_t* token);
+OAK_API int oak_token_size(const struct oak_token_t* token);
 OAK_API const char* oak_token_text(const struct oak_token_t* token);
 OAK_API int oak_token_as_i32(const struct oak_token_t* token);
 OAK_API float oak_token_as_f32(const struct oak_token_t* token);
