+++ src/common/oak_module_loader_path.c
@@ -109,7 +109,7 @@ char* dotted_name_from_path(struct oak_allocator_t* a,
         oak_container_of(pos, struct oak_ast_node_t, link);
     if (!first)
       buf[w++] = '.';
-    const usize len = oak_token_size(ident->token);
+    const int len = oak_token_size(ident->token);
     memcpy(buf + w, oak_token_text(ident->token), len);
     w += len;
     first = 0;
